requirements.txt
streamlit
google-generativeai
python-dotenv
pymupdf